feat: Add functionality to download files from XML
Refactor the code to add the functionality of downloading files from XML. This includes adding the `downloader.py` and `file_manager.py` modules. The `downloader.py` module contains the `download_files` function that uses multithreading to download files concurrently. The `file_manager.py` module fetches the URLs from the XML file and returns a list of URLs to be downloaded. This commit also includes refactoring the `main.py` script to use the new modules and fetch the URLs from the XML file.